requirements.txt
mysql-connector-python==8.3.0
bcrypt==4.1.2
Pillow==10.2.0  # For image handling in Tkinter
python-dotenv==1.0.1  # For environment variables
